+++ CHANGELOG.md
@@ -16,6 +16,8 @@ prior to assuming the existence of said check.
 - Additional logging around handlers
 - Accept additional time formats in sensuctl
 - Entities can now be created via sensuctl.
+- Added the format `wrapped-json` to sensuctl `configure`, `list` and `info`
+commands, which is compatible with `sensuctl create`.
 
 ### Changed
 - Add logging around the Sensu event pipeline.